requirements.txt
python-dotenv
loguru
langgraph
langchain
numpy
pyautogui
Pillow
pywin32
fastapi
mss
azure-cognitiveservices-vision-computervision
opencv-python
langchain_anthropic
langchain_openai
langchain_google_genai
langchain_xai
langchain_deepseek
pinecone
pinecone-client
py-zerox